feat(hover): add text and attribution interpolation and DRY
